feat: tests and final rest api

+++ internal/domain/user.go
@@ -5,15 +5,17 @@ type User struct {
 	ID        uint
 	Email     string
 	Password  string
-	CreatedAt int
-	UpdatedAt int
+	CreatedAt int64
+	UpdatedAt int64
 }
 
+// model of authorisation request
 type UserAuthReq struct {
 	Email    string
 	Password string
 }
 
+// model of registration request
 type UserRegisterReq struct {
 	Email      string
 	Password   string
